move candidate_from_obligation_no_cache
